Fixed save location for show files. Ensured that projects are delted when changing project types. Fixed bug if a filesystem project is delete but does not exist causing an exception to be thrown.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@4300 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/Controller/Editor/ProjectSupport/FilesystemProjectType.cs
@@ -10,7 +10,10 @@ namespace Medical.Editor
     {
         public void deleteProject(string name)
         {
-            Directory.Delete(name, true);
+            if (Directory.Exists(name))
+            {
+                Directory.Delete(name, true);
+            }
         }
 
         public void ensureProjectExists(string name)
